+++ R/learner_dbarts_regr_bart.R
@@ -9,19 +9,16 @@
 #' @template learner
 #' @templateVar id regr.bart
 #'
-#' @section Initial parameter values:
+#' @section Custom mlr3 parameters:
 #' * Parameter: offset
 #'   * The parameter is removed, because only `dbarts::bart2` allows an offset during training,
 #'     and therefore the offset parameter in `dbarts:::predict.bart` is irrelevant for
 #'     `dbarts::dbart`.
-#' * Parameter: nthread, nchain, combineChains, combinechains
+#' * Parameter: nchain, combineChains, combinechains
 #'   * The parameters are removed as parallelization of multiple models is handled by future.
 #'
-#' @section Custom mlr3 defaults:
-#' * Parameter: keeptrees
-#'   * Original: FALSE
-#'   * New: TRUE
-#'   * Reason: Required for prediction
+#' @section Initial parameter values:
+#' * `keeptrees` is initialized to `TRUE` because it is required for prediction.
 #'
 #' @references
 #' `r format_bib("sparapani2021nonparametric", "chipman2010bart")`
@@ -53,13 +50,15 @@ LearnerRegrBart = R6Class("LearnerRegrBart",
         usequants = p_lgl(default = FALSE, tags = "train"),
         numcut = p_int(default = 100L, lower = 1L, tags = "train"),
         printcutoffs = p_int(default = 0, tags = "train"),
-        verbose = p_lgl(default = TRUE, tags = "train"),
+        verbose = p_lgl(default = FALSE, tags = "train"),
+        nthread = p_int(default = 1L, tags = c("train", "threads")),
         keeptrees = p_lgl(default = FALSE, tags = "train"),
         keepcall = p_lgl(default = TRUE, tags = "train"),
         sampleronly = p_lgl(default = FALSE, tags = "train"),
         seed = p_int(default = NA_integer_, tags = "train", special_vals = list(NA_integer_)),
         proposalprobs = p_uty(default = NULL, tags = "train"),
-        splitprobs = p_uty(default = NULL, tags = "train")
+        splitprobs = p_uty(default = NULL, tags = "train"),
+        keepsampler = p_lgl(default = NO_DEF, tags = "train")
       )
 
       ps$values = list(keeptrees = TRUE)
@@ -69,7 +68,7 @@ LearnerRegrBart = R6Class("LearnerRegrBart",
         packages = c("mlr3extralearners", "dbarts"),
         feature_types = c("integer", "numeric", "factor", "ordered"),
         # TODO: add "se" to the list of predict types.
-        predict_types = c("response"),
+        predict_types = "response",
         param_set = ps,
         properties = c("weights"),
         man = "mlr3extralearners::mlr_learners_regr.bart",