+++ R/learner_lightgbm_classif_lightgbm.R
@@ -14,27 +14,27 @@
 #' @templateVar id classif.lightgbm
 #'
 #' @section Initial parameter values:
+#' * `num_threads`:
+#'   * Actual default: 0L
+#'   * Initial value: 1L
+#'   * Reason for change: Prevents accidental conflicts with `future`.
+#' * `verbose`:
+#'   * Actual default: 1L
+#'   * Initial value: -1L
+#'   * Reason for change: Prevents accidental conflicts with mlr messaging system.
+#' @section Custom mlr3 defaults:
+#' * `objective`:
+#'   Depending if the task is binary / multiclass, the default is `"binary"` or
+#'   `"multiclasss"`.
+#' @section Custom mlr3 parameters:
+#' * `early_stopping`
+#'   Whether to use the test set for early stopping. Default is `FALSE`.
 #' * `convert_categorical`:
 #'   Additional parameter. If this parameter is set to `TRUE` (default), all factor and logical
 #'   columns are converted to integers and the parameter categorical_feature of lightgbm is set to
 #'   those columns.
 #' * `num_class`:
 #'  This parameter is automatically inferred for multiclass tasks and does not have to be set.
-#' @section Custom mlr3 defaults:
-#' * `num_threads`:
-#'   * Actual default: 0L
-#'   * Adjusted default: 1L
-#'   * Reason for change: Prevents accidental conflicts with `future`.
-#' * `verbose`:
-#'   * Actual default: 1L
-#'   * Adjusted default: -1L
-#'   * Reason for change: Prevents accidental conflicts with mlr messaging system.
-#' * `objective`:
-#'   Depending if the task is binary / multiclass, the default is set to `"binary"` or
-#'   `"multiclasss"`.
-#' @section Custom mlr3 parameters:
-#' * `early_stopping`
-#'   Whether to use the test set for early stopping. Default is `FALSE`.
 #'
 #' @references
 #' `r format_bib("ke2017lightgbm")`
@@ -127,6 +127,10 @@ LearnerClassifLightGBM = R6Class("LearnerClassifLightGBM",
         cegb_penalty_feature_coupled = p_uty(tags = "train"),
         path_smooth = p_dbl(default = 0.0, lower = 0.0, tags = "train"),
         interaction_constraints = p_uty(tags = "train"),
+        use_quantized_grad = p_lgl(default = TRUE, tags = "train"),
+        num_grad_quant_bins = p_int(default = 4, tags = "train"),
+        quant_train_renew_leaf = p_lgl(default = FALSE, tags = "train"),
+        stochastic_rounding = p_lgl(default = TRUE, tags = "train"),
 
         # dataset parameters
         max_bin = p_int(default = 255L, lower = 2L, tags = "train"),