+++ R/learner_mboost_regr_glmboost.R
@@ -50,7 +50,7 @@ LearnerRegrGLMBoost = R6Class("LearnerRegrGLMBoost",
         id = "regr.glmboost",
         packages = c("mlr3extralearners", "mboost"),
         feature_types = c("integer", "numeric", "factor", "ordered"),
-        predict_types = c("response"),
+        predict_types = "response",
         param_set = ps,
         properties = "weights",
         man = "mlr3extralearners::mlr_learners_regr.glmboost",
@@ -82,7 +82,7 @@ LearnerRegrGLMBoost = R6Class("LearnerRegrGLMBoost",
 
       if (self$param_set$values$family != "custom") {
         pars_family = pars[which(names(pars) %in%
-          formalArgs(utils::getFromNamespace(
+          formalArgs(getFromNamespace(
             pars_glmboost$family,
             asNamespace("mboost"))))]
       }