+++ R/learner_obliqueRSF_surv_obliqueRSF.R
@@ -10,17 +10,14 @@
 #' @template learner
 #' @templateVar id surv.obliqueRSF
 #'
-#' @section Initial parameter values:
+#' @section Custom mlr3 parameters:
 #' - `mtry`:
 #'   - This hyperparameter can alternatively be set via the added hyperparameter `mtry_ratio`
 #'     as `mtry = max(ceiling(mtry_ratio * n_features), 1)`.
 #'     Note that `mtry` and `mtry_ratio` are mutually exclusive.
 #'
-#' @section Custom mlr3 defaults:
-#' - `verbose`:
-#'   - Actual default: `TRUE`
-#'   - Adjusted default: `FALSE`
-#'   - Reason for change: mlr3 already has it's own verbose set to `TRUE` by default
+#' @section Initial parameter values:
+#' - `verbose` is initialized to `FALSE`
 #'
 #' @references
 #' `r format_bib("jaeger_2019")`