+++ tests/testthat/test_paramtest_dbarts_classif_bart.R
@@ -6,14 +6,14 @@ test_that("paramtest classif.bart train", {
     "y.train", # handled internally
     "x.test", # handled internally
     "weights", # handled by task
-    "nthread", # handled by future
     "nchain", # unused (specifies how many independent tree sets and fits should be calculated
     # This can be done with future
     "combinechains", # only used if nchain > 1
     "sigest", # regr only
     "sigdf", # regr only
     "sigquant", # regr only
-    "keeptrees" # must be TRUE for prediction
+    "keeptrees", # must be TRUE for prediction
+    "weights" # train predict weights not clear from the documentation
   )
 
   paramtest = run_paramtest(learner, fun_list, exclude, tag = "train")
@@ -28,7 +28,8 @@ test_that("paramtest classif.bart predict", {
     "newdata", # handled internally
     "object", # handled internally
     "type", # only allow the response
-    "offset" # we don't allow offsets (only using bart2 allows to train with offset, theoretically
+    "offset", # we don't allow offsets (only using bart2 allows to train with offset, theoretically
+    "weights" # train predict weights not clear from the documentation
     # the predict function still uses an offset but this will only be implemented if a user demands
     # it (https://cran.r-project.org/web/packages/dbarts/dbarts.pdf)
   )