+++ tests/testthat/test_paramtest_lightgbm_classif_lightgbm.R
@@ -80,7 +80,10 @@ test_that("paramtest classif.lightgbm train", {
     "valids", # handled internally
     "obj", # alias for objective
     "colnames", # we don't want to allow overwriting any names
-    "init_model" # handled internally via hotstarting
+    "init_model", # handled internally via hotstarting
+
+
+    "lambdarank_position_bias_regularization" # only for lambdarank
 
   )