+++ tests/testthat/test_paramtest_mgcv_classif_gam.R
@@ -3,6 +3,7 @@ test_that("classif.gam train", {
   fun_list = list(mgcv::gam, mgcv::gam.control)
 
   exclude = c(
+    "outerPIsteps", # weird errors in rcmdcheck (different result on devel and release)
     "family", # handled relatively to type of task
     "data", # handled internally
     "weights", # handled internally