+++ tests/testthat/test_paramtest_randomForestSRC_classif_rfsrc.R
@@ -9,8 +9,7 @@ test_that("paramtest classif.rfsrc train", {
     "case.wt", # handled by task weights
     "mtry.ratio",
     "sampsize.ratio",
-    "cores", # is set as option(rf.cores)
-    "perf.type" # scoring is done with measures
+    "cores" # is set as option(rf.cores)
   )
 
   paramtest = run_paramtest(learner, fun_list, exclude, tag = "train")